requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.9
requests==2.31.0
openai==0.27.8
python-dotenv==1.0.0
Pillow==9.4.0
pdf2image==1.16.3
langchain
pypdf
python-multipart